feat: remove intermediate original sdk configuration api protocol & update tests
