+++ PrebidMobileTests/TargetingObjCTests.m
@@ -272,8 +272,8 @@ limitations under the License.
 
 - (void)testAccessControlList {
     //given
-    NSString *bidderNameRubicon = Prebid.bidderNameRubiconProject;
-    NSString *bidderNameAppNexus = Prebid.bidderNameAppNexus;
+    NSString *bidderNameRubicon = PrebidConfiguration.bidderNameRubiconProject;
+    NSString *bidderNameAppNexus = PrebidConfiguration.bidderNameAppNexus;
     
     //when
     [Targeting.shared addBidderToAccessControlList:bidderNameRubicon];
